+++ tests/testthat/test-bootstrap.R
@@ -55,6 +55,8 @@ bootstrap <- function(df, func, R) {
 
 test_that("Small Poisson bootstrap examples with and without weights", {
 
+    # skip_on_cran()
+
     set.seed(123)
 
     dobson <- data.frame(
@@ -104,6 +106,9 @@ test_that("Small Poisson bootstrap examples with and without weights", {
 })
 
 test_that("Small Poisson bootstrap examples with an offset", {
+
+    # skip_on_cran()
+
     dobson <- data.frame(
         counts = c(18,17,15,20,10,20,25,13,12),
         outcome2 = c(0,1,0,0,1,0,0,1,0),
@@ -132,31 +137,7 @@ test_that("Small Poisson bootstrap examples with an offset", {
     expect_equal(cbo$summary$bpi_lower + c(1,0,0,0,0), cbn$summary$bpi_lower, tolerance = 1E-4)
 })
 
-# test_that("Large Cox with weights and bootstrapping", {
-#     set.seed(123)
-#     sim <- simulateCyclopsData(nstrata=1000,
-#                                ncovars=10,
-#                                nrows=10000,
-#                                effectSizeSd=0.5,
-#                                eCovarsPerRow=2,
-#                                model="survival")
-#     sim$outcomes$weights <- 1/sim$outcomes$rr
-#
-#     # Cyclops
-#     cyclopsData <- convertToCyclopsData(outcomes = sim$outcomes,
-#                                         covariates = sim$covariates,
-#                                         modelType = "cox")
-#     fitCyclops <- fitCyclopsModel(cyclopsData = cyclopsData)
-#
-#     bs <- runBootstrap(fitCyclops, outFileName = "out.txt", treatmentId = "1", replicates = 100)
-#     result <- read.csv("out.txt")
-#     result
-# })
 
-# empinf(bb)
-#
-# mat <- matrix(nrow = Cyclops::getNumberOfRows(cd))
-#
 # boot.out <- list(
 #     t = as.matrix(cb$samples),
 #     sim = "ordinary",
@@ -170,6 +151,9 @@ test_that("Small Poisson bootstrap examples with an offset", {
 # boot.ci(boot.out, index = 1L, type = "bca")
 
 test_that("bootstrap option for na.rm", {
+
+    # skip_on_cran() # ASAN error https://github.com/OHDSI/Cyclops/actions/runs/16373844416
+
     test <- read.table(header=T, sep = ",", text = "
 start, length, event, x1, x2
 0, 4,  1,0,0
@@ -199,6 +183,9 @@ start, length, event, x1, x2
 })
 
 test_that("Large logistic bootstrap with and without weights", {
+
+    # skip_on_cran()
+
     set.seed(123)
     sim <- simulateCyclopsData(nstrata=100,
                                ncovars=4,
@@ -223,11 +210,11 @@ test_that("Large logistic bootstrap with and without weights", {
         }, R = 1999)
 
     bbNoStdError <- sqrt(apply(bbNoWeights, 2L, var))
-    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.01)
+    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.02)
 
     expect_equivalent(bsNoWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbNoWeights),
-                      tolerance = 0.05)
+                      tolerance = 0.075)
 
     sim$outcomes$weights <- rep(c(0.1,0.9), nrow(sim$outcomes) / 2)
 
@@ -248,7 +235,7 @@ test_that("Large logistic bootstrap with and without weights", {
         }, R = 1999)
 
     bbYesStdError <- sqrt(apply(bbYesWeights, 2L, var))
-    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.01)
+    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.02)
 
     expect_equivalent(bsYesWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbYesWeights),
@@ -256,6 +243,9 @@ test_that("Large logistic bootstrap with and without weights", {
 })
 
 test_that("Large Poisson bootstrap with and without weights", {
+
+    # skip_on_cran()
+
     set.seed(123)
     sim <- simulateCyclopsData(nstrata=100,
                                ncovars=4,
@@ -280,11 +270,11 @@ test_that("Large Poisson bootstrap with and without weights", {
         }, R = 1999)
 
     bbNoStdError <- sqrt(apply(bbNoWeights, 2L, var))
-    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.01)
+    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.02)
 
     expect_equivalent(bsNoWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbNoWeights),
-                      tolerance = 0.05)
+                      tolerance = 0.075)
 
     sim$outcomes$weights <- rep(c(0.1,0.9), nrow(sim$outcomes) / 2)
 
@@ -305,14 +295,17 @@ test_that("Large Poisson bootstrap with and without weights", {
         }, R = 1999)
 
     bbYesStdError <- sqrt(apply(bbYesWeights, 2L, var))
-    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.01)
+    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.02)
 
     expect_equivalent(bsYesWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbYesWeights),
-                      tolerance = 0.05)
+                      tolerance = 0.075)
 })
 
 test_that("Large Cox bootstrap with and without weights", {
+
+    # skip_on_cran()
+
     set.seed(123)
     sim <- simulateCyclopsData(nstrata=100,
                                ncovars=4,
@@ -339,11 +332,11 @@ test_that("Large Cox bootstrap with and without weights", {
         }, R = 3999)
 
     bbNoStdError <- sqrt(apply(bbNoWeights, 2L, var))
-    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.01)
+    expect_equivalent(bsNoWeights$summary$std_err, bbNoStdError, tolerance = 0.02)
 
     expect_equivalent(bsNoWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbNoWeights),
-                      tolerance = 0.05)
+                      tolerance = 0.075)
 
     sim$outcomes$weights <- rep(c(0.1,0.9), nrow(sim$outcomes) / 2)
 
@@ -364,9 +357,9 @@ test_that("Large Cox bootstrap with and without weights", {
         }, R = 3999)
 
     bbYesStdError <- sqrt(apply(bbYesWeights, 2L, var))
-    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.01)
+    expect_equivalent(bsYesWeights$summary$std_err, bbYesStdError, tolerance = 0.02)
 
     expect_equivalent(bsYesWeights$summary[,c("bpi_lower", "bpi_upper")],
                       getAllPercentileIntervals(bbYesWeights),
-                      tolerance = 0.05)
+                      tolerance = 0.075)
 })